requirements.txt
requests
pandas
matplotlib
numpy
# smtplib
python-dotenv